requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
botocore==1.24.38
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
gql==3.2.0
graphql-core==3.2.1
idna==3.3
jmespath==1.0.0
multidict==6.0.2
pyTelegramBotAPI==4.4.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
websockets==10.2
yarl==1.7.2
